Renamed PhpstanCheckDeprecationTask to PhpStanTask
