+++ config/grumphp.yml
@@ -8,7 +8,7 @@ grumphp:
     check_file_permissions: ~
     php_check_syntax: ~
     phpcs: ~
-    phpstan_check_deprecation: ~
+    php_stan: ~
     ecs: ~
     yaml_lint: ~
     json_lint: ~
@@ -18,6 +18,6 @@ grumphp:
     - Wunderio\GrumPHP\Task\CheckFilePermissions\CheckFilePermissionsExtensionLoader
     - Wunderio\GrumPHP\Task\Ecs\EcsExtensionLoader
     - Wunderio\GrumPHP\Task\Phpcs\PhpcsExtensionLoader
-    - Wunderio\GrumPHP\Task\PhpstanCheckDeprecation\PhpstanCheckDeprecationExtensionLoader
+    - Wunderio\GrumPHP\Task\PhpStan\PhpStanExtensionLoader
     - Wunderio\GrumPHP\Task\YamlLint\YamlLintExtensionLoader
     - Wunderio\GrumPHP\Task\JsonLint\JsonLintExtensionLoader
