+++ src/Task/PhpStan/PhpStanExtensionLoader.php
@@ -1,0 +1,14 @@
+<?php
+
+namespace Wunderio\GrumPHP\Task\PhpStan;
+
+use Wunderio\GrumPHP\Task\AbstractExternalExtensionLoader;
+
+/**
+ * Class PhpStan.
+ *
+ * Register PhpStan task in the service container of GrumPHP.
+ *
+ * @package Wunderio\GrumPHP\Task
+ */
+class PhpStanExtensionLoader extends AbstractExternalExtensionLoader {}
