--- src/Task/PhpstanCheckDeprecation/PhpstanCheckDeprecationExtensionLoader.php
@@ -1,14 +1,0 @@
-<?php
-
-namespace Wunderio\GrumPHP\Task\PhpstanCheckDeprecation;
-
-use Wunderio\GrumPHP\Task\AbstractExternalExtensionLoader;
-
-/**
- * Class PhpstanCheckDeprecationExtensionLoader.
- *
- * Register PhpstanCheckDeprecation task in the service container of GrumPHP.
- *
- * @package Wunderio\GrumPHP\Task
- */
-class PhpstanCheckDeprecationExtensionLoader extends AbstractExternalExtensionLoader {}
